Ennemy factory is working, you can now launch a wave from the right gui. (a short desc of the next incoming wave is also displayed on the right).

+++ cafard.h
@@ -10,7 +10,7 @@ class Map;
 class Cafard : public Monster
 {
 public:
-    Cafard(Map* map,int posx,int posy, int size);
+    Cafard(Map* map,int posx,int posy, float size);
 
 protected:
     void advance(int phase);
