+++ monster.cpp
@@ -1,6 +1,6 @@
 #include "monster.h"
 
-Monster::Monster(int posx,int posy, int size):Entity(posx,posy),_size(size){ }
+Monster::Monster(int posx,int posy, float size):Entity(posx,posy),_size(size){ }
 
 
 int Monster::getHP(void) const { return _hp; }
